+++ changelog.md
@@ -7,6 +7,8 @@
 - Updated koin version to 3.0.1
 - Removed jCenter dependency and added aar files temporarily in libs folder
 - Verify BPay
+- Payment Limits APIs
+- Add new payment error codes
 
 ### 3.9.0